+++ dev4press/core/options/Process.php
@@ -63,16 +63,12 @@ class Process {
 		return $this;
 	}
 
-	public function process( $request = false ) : array {
+	public function process( array $request ) : array {
 		$list = array();
 
 		foreach ( $this->settings as $setting ) {
 			if ( $setting->type != '_' ) {
-				if ( $request === false ) {
-					$post = $_REQUEST[ $this->base ][ $setting->type ] ?? array(); // phpcs:ignore WordPress.Security.NonceVerification,WordPress.Security.ValidatedSanitizedInput
-				} else {
-					$post = $request[ $this->base ][ $setting->type ] ?? array();
-				}
+				$post = $request[ $this->base ][ $setting->type ] ?? array();
 
 				$list[ $setting->type ][ $setting->name ] = $this->process_single( $setting, $post );
 			}